requirements.txt
streamlit
paho-mqtt
pandas
sqlalchemy
pymysql
